Convert all loggers to typescript
