+++ logger/compact-logger.ts
@@ -1,38 +1,51 @@
 'use strict';
 
-const path = require( 'path' );
+// const path = require( 'path' );
+// const chalk = require( 'chalk' );
+// const figures = require( 'figures' );
+// const log = require( 'log-update' );
+// const ProgressPlugin = require( 'webpack/lib/ProgressPlugin' );
 
-const chalk = require( 'chalk' );
-const figures = require( 'figures' );
-
-const ProgressPlugin = require( 'webpack/lib/ProgressPlugin' );
+import path from 'path';
+import chalk from 'chalk';
+import figures from 'figures';
+import log from 'log-update';
+import { ProgressPlugin } from 'webpack';
+import { InternalOptions } from '..';
 
 /**
- * Expanded Logger
+ * Compact Logger
  */
-module.exports = function ExpandedLogger( options ) {
+export default function CompactLogger( options: InternalOptions ) {
 
 	// Configure color
 	chalk.enabled = options.color;
 
-	const absoluteProjectPath = `${ path.resolve( '.' ).toString() }`;
+	const absoluteProjectPath: string = `${ path.resolve( '.' ).toString() }`;
 
 	// Variables for the process, reset after each run
-	let startTime;
-	let previousStep = 0;
+	let startTime: number;
+	let previousStep: number = 0;
 
 	// Initial log
-	console.log( chalk.white( 'Webpack: Starting ...' ) );
+	let logLines: string[] = [];
+	log( 'Webpack: Starting ...' );
 
 	/**
 	 * Use the webpack-internal progress plugin as the base of the logger
 	 */
 	return new ProgressPlugin( ( progress, message, moduleProgress, activeModules, moduleName ) => {
 
+		// Reset log output
+		logLines = [];
+
 		// Reset process variables for this run
 		if ( previousStep === 0 ) {
 			startTime = new Date().getTime();
 		}
+
+		// STEP 0: HEADER
+		logLines.push( chalk.white( 'Webpack: Starting ...\n' ) );
 
 		// STEP 1: COMPILATION
 		if ( progress >= 0 && progress < 0.1 ) {
@@ -40,10 +53,14 @@ module.exports = function ExpandedLogger( options ) {
 			// Skip if we jumped back a step, else update the step counter
 			if ( previousStep > 1 ) {
 				return;
-			} else if ( previousStep < 1 ) {
-				console.log( chalk.white( `\n  ${ figures.pointer } Compile modules` ) );
 			}
 			previousStep = 1;
+
+			logLines.push( chalk.white( `  ${ figures.pointer } Compile modules` ) );
+
+		} else if ( progress >= 0.1 ) {
+
+			logLines.push( chalk.green( `  ${ figures.tick } Compile modules` ) );
 
 		}
 
@@ -53,15 +70,16 @@ module.exports = function ExpandedLogger( options ) {
 			// Skip if we jumped back a step, else update the step counter
 			if ( previousStep > 2 ) {
 				return;
-			} else if ( previousStep < 2 ) {
-				console.log( chalk.white( `\n  ${ figures.pointer } Build modules` ) );
 			}
 			previousStep = 2;
+
+			// Log progress line (with sub-progress indicator)
+			const subProgress = Math.round( ( progress - 0.1 ) * 10000 / 60 );
+			logLines.push( chalk.white( `  ${ figures.pointer } Build modules (${ subProgress }%)` ) );
 
 			// Log additional information (if possible)
 			if ( moduleName !== undefined ) {
 
-				const roundedSubProgress = Math.round( ( progress - 0.1 ) * 10000 / 60 );
 				let betterModuleName = moduleName;
 
 				// Only show the file that is actually being processed (and remove all details about used loaders)
@@ -92,10 +110,14 @@ module.exports = function ExpandedLogger( options ) {
 				}
 
 				const [ betterModulesDone, betterAllModules ] = moduleProgress.split( '/' );
-				const moduleDetails = `${ betterModulesDone } of ${ betterAllModules } :: ${ betterModuleName }`;
-				console.log( chalk.grey( `    ${ figures.arrowRight } [${ roundedSubProgress }%] ${ moduleDetails }` ) );
+				const moduleDetails: string = `${ betterModulesDone } of ${ betterAllModules } :: ${ betterModuleName }`;
+				logLines.push( chalk.grey( `    ${ figures.arrowRight } ${ moduleDetails }` ) );
 
 			}
+
+		} else if ( progress > 0.7 ) {
+
+			logLines.push( chalk.green( `  ${ figures.tick } Build modules` ) );
 
 		}
 
@@ -105,18 +127,21 @@ module.exports = function ExpandedLogger( options ) {
 			// Skip if we jumped back a step, else update the step counter
 			if ( previousStep > 3 ) {
 				return;
-			} else if ( previousStep < 3 ) {
-				console.log( chalk.white( `\n  ${ figures.pointer } Optimize modules` ) );
 			}
 			previousStep = 3;
 
 			// Log progress line (with sub-progress indicator)
-			const subProgress = Math.round( ( progress - 0.71 ) * 10000 / 23 );
+			const subProgress: number = Math.round( ( progress - 0.71 ) * 10000 / 23 );
+			logLines.push( chalk.white( `  ${ figures.pointer } Optimize modules (${ subProgress }%)` ) );
 
-			const formattedMessage = `${ message[ 0 ].toUpperCase() }${ message.slice( 1 ) }`;
-			const formattedMessageExtra = progress === 0.91 ? ' -- may take a while' : ''; // Add some extra info (calming devs down)
+			const formattedMessage: string = `${ message[ 0 ].toUpperCase() }${ message.slice( 1 ) }`;
+			const formattedMessageExtra: string = progress === 0.91 ? ' -- may take a while' : ''; // Add some extra info (calming devs down)
 
-			console.log( chalk.grey( `    ${ figures.arrowRight } [${ subProgress }%] ${ formattedMessage }${ formattedMessageExtra } ...` ) );
+			logLines.push( chalk.grey( `    ${ figures.arrowRight } ${ formattedMessage }${ formattedMessageExtra } ...` ) );
+
+		} else if ( progress >= 0.95 ) {
+
+			logLines.push( chalk.green( `  ${ figures.tick } Optimize modules` ) );
 
 		}
 
@@ -126,10 +151,14 @@ module.exports = function ExpandedLogger( options ) {
 			// Skip if we jumped back a step, else update the step counter
 			if ( previousStep > 4 ) {
 				return;
-			} else if ( previousStep < 4 ) {
-				console.log( chalk.white( `\n  ${ figures.pointer } Emit files` ) );
 			}
 			previousStep = 4;
+
+			logLines.push( chalk.white( `  ${ figures.pointer } Emit files` ) );
+
+		} else if ( progress === 1 ) {
+
+			logLines.push( chalk.green( `  ${ figures.tick } Emit files` ) );
 
 		}
 
@@ -138,11 +167,17 @@ module.exports = function ExpandedLogger( options ) {
 
 			// Calculate process time
 			previousStep = 0;
-			const finishTime = new Date().getTime();
-			const processTime = ( ( finishTime - startTime ) / 1000 ).toFixed( 3 );
+			const finishTime: number = new Date().getTime();
+			const processTime: string = ( ( finishTime - startTime ) / 1000 ).toFixed( 3 );
 
-			console.log( chalk.white( `\nWebpack: Finished after ${ processTime } seconds.\n` ) );
+			logLines.push( chalk.white( `\nWebpack: Finished after ${ processTime } seconds.\n` ) );
 
+		}
+
+		// Finally, let's bring those logs to da screen
+		log( logLines.join( '\n' ) );
+		if ( progress === 1 ) {
+			log.done();
 		}
 
 	} );
